Rcpp pw and dpw
Best implementation of Rcpp on pw and dpw I could figure.  It is still not in the package, you will have to see the file "calculate_pw_test.R" in the scratch folder.  The other file "calculate_pw_test2.R" contains an attempt a full C implementation.  This did not have much improvement over the first, and I do not recommend using it.